Add keybaoard fuc about show and hide
In the swift, if developer want controll keyboard make decision
when keyboard show and hied. so I add func aout that. in this commit in loginviewcontroller extention funtion named keyboardwillshow and keyboardheight both of them impliment keyboard. it's not perfect yet but I add some more function to help this implimentation.